// source/frontend/src/models/api/generated/ApiGen_Concepts_Notification.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { UtcIsoDate } from '@/models/api/UtcIsoDateTime';

import { ApiGen_Base_BaseConcurrent } from './ApiGen_Base_BaseConcurrent';

// LINK: @backend/apimodels/Models/Concepts/Notification/NotificationModel.cs
export interface ApiGen_Concepts_Notification extends ApiGen_Base_BaseConcurrent {
  notificationId: number | null;
  notificationTypeCode: string | null;
  notificationTriggerDate: UtcIsoDate;
  notificationMessage: string | null;
  acquisitionFileId: number | null;
  dispositionFileId: number | null;
  researchFileId: number | null;
  managementFileId: number | null;
  leaseId: number | null;
  takeId: number | null;
  insuranceId: number | null;
  leaseConsultationId: number | null;
  noticeOfClaimId: number | null;
  leaseRenewalId: number | null;
  expropOwnerHistoryId: number | null;
  agreementId: number | null;
}